requirements.txt
slackclient
python-dotenv
Flask
slackeventsapi
g4f
slack_bolt
slack_sdk

langchain
playwright
unstructured